// src/models/ATM.ts
export interface ATM {
    atm: string;
    sitio: string;
    cr: string;
    divison: string;
    marca: string;
    tipo_dispositivo: string;
    estatus_dispositivo: string;
    calle: string;
    num_ext_: string;
    estado: string;
    ciudad: string;
    cp: string;
    del_muni: string;
    colonia: string;
    latitud: string;
    longitud: string;
    tipo_localidad: string;
    IDC: string;
    ETV: string;
}